feature/KAN-18 renamed PasswordVisibility component to PasswordVisibilityButton
