+++ src/components/passwordVisibility/__test__/PasswordVisibilityButton.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import { vi } from "vitest";
 import { TEST_ID } from "./testIds";
 
-import PasswordVisibility from "../PasswordVisibility";
+import PasswordVisibility from "../PasswordVisibilityButton";
 import userEvent from "@testing-library/user-event";
 
 describe("Password visibility component", () => {
